backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
pydantic>=2.11.0
pydantic-settings>=2.11.0
openai-agents>=0.9.0
